project_requirements.txt
flask==2.3.3
gunicorn==23.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-telegram-bot==20.7
numpy==1.26.3
pillow==10.1.0
flask-login==0.6.3
werkzeug==2.3.7
flask-wtf==1.2.1
email-validator==2.1.0
python-dotenv==1.0.0
flask-apscheduler==1.13.1
requests==2.32.3
scikit-image==0.22.0